moved all discretization classes in one file - discretize.py
